functions/oss_callback/backend/requirements.txt
tensorflow==1.13.1
nltk==3.5
pandas==1.0.3
tqdm==4.52.0
pillow==8.0.1
opencv-python==4.4.0.46
oss2==2.13.1
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-alimt==3.0.28
bottle==0.12.19
gensim==3.8.3
jieba==0.42.1
numpy==1.18.1
scipy==1.4.1